% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2.R
\name{arc2}
\alias{arc2}
\title{Arc2 - Africa Rainfall Climatology version 2}
\usage{
arc2(date, ...)
}
\arguments{
\item{date}{(character/date) one or more dates of the form YYYY-MM-DD}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a tibble/data.frame with columns:
\itemize{
\item date: date (YYYY-MM-DD)
\item lon: longitude
\item lat: latitude
\item precip: precipitation (mm)
}
}
\description{
Arc2 - Africa Rainfall Climatology version 2
}
\note{
See \link{arc2_cache} for managing cached files
}
\examples{
\dontrun{
x = arc2(date = "1983-01-01")
arc2(date = "2017-02-14")

# many dates
arc2(date = c("2019-05-27", "2019-05-28"))
arc2(seq(as.Date("2019-04-21"), by = "day", length.out = 5))
## combine outputs 
x <- arc2(seq(as.Date("2019-05-20"), as.Date("2019-05-25"), "days"))
dplyr::bind_rows(x)
}
}
\references{
docs:
\url{ftp://ftp.cpc.ncep.noaa.gov/fews/fewsdata/africa/arc2/ARC2_readme.txt}
}
