% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{argo}
\alias{argo}
\alias{argo_files}
\alias{argo_search}
\title{Get Argo buoy data}
\usage{
argo_search(gt = NULL, of = NULL, qwmo = NULL, wmo = NULL, box = NULL,
  limit = 10, ...)

argo_files(gt = NULL, of = NULL, qwmo = NULL, wmo = NULL, box = NULL,
  ...)
}
\arguments{
\item{limit}{(integer) number to return}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}. Optional}
}
\description{
Get Argo buoy data
}
\examples{
\dontrun{
argo_search("n", limit = 3)
argo_search("np", limit = 3)
argo_search("nf", limit = 3)
argo_search("coord", limit = 3)
argo_search("fullcoord", limit = 5)
argo_search("list", "dac", limit = 5)
argo_search(qwmo = 49, limit = 3)
argo_search("n", box = c(-40, 35, 3, 2))

argo_files(wmo = 4900881)
}
}
\references{
\url{http://www.ifremer.fr/lpo/naarc/m/docs/api/howto.html}
}

