% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-autoplot.R
\name{autoplot.meteo_coverage}
\alias{autoplot.meteo_coverage}
\title{autoplot method for meteo_coverage objects}
\usage{
\method{autoplot}{meteo_coverage}(mateo_coverage, old_style = FALSE)
}
\arguments{
\item{mateo_coverage}{the object returned from \code{\link[=meteo_coverage]{meteo_coverage()}}}

\item{old_style}{(logical) create the old style of plots, which is faster, but
does not plot gaps to indicate missing data}
}
\value{
A ggplot2 plot
}
\description{
autoplot method for meteo_coverage objects
}
\details{
see \code{\link[=meteo_coverage]{meteo_coverage()}} for examples
}
