% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-autoplot.R
\name{autoplot_meteo_coverage}
\alias{autoplot_meteo_coverage}
\title{autoplot method for meteo_coverage objects}
\usage{
autoplot_meteo_coverage(meteo_object, old_style = FALSE)
}
\arguments{
\item{meteo_object}{the object returned from \code{\link[=meteo_coverage]{meteo_coverage()}}}

\item{old_style}{(logical) create the old style of plots, which is faster, but
does not plot gaps to indicate missing data}
}
\value{
A ggplot2 plot
}
\description{
autoplot method for meteo_coverage objects
}
\details{
see \code{\link[=meteo_coverage]{meteo_coverage()}} for examples.
}
