% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buoy}
\alias{buoy}
\alias{print.buoy}
\title{Get NOAA buoy data from the National Buoy Data Center}
\usage{
buoy(dataset = NULL, buoyid = NULL, datatype = NULL, year = NULL, ...)

\method{print}{buoy}(x, ...)
}
\arguments{
\item{dataset}{Dataset to query. See below for details. (required)}

\item{buoyid}{Buoy id. (optional)}

\item{datatype}{Data type, one of 'c', 'cc', 'p'. (optional)}

\item{year}{Year of data collection. (optional)}

\item{...}{Further arguments passed on to the API GET call. (optional)}

\item{x}{Input to print, output from buoy function.}
}
\value{
A data.frame
}
\description{
Get NOAA buoy data from the National Buoy Data Center
}
\details{
Options for the dataset parameter. One of:
\itemize{
 \item adcp - Acoustic Doppler Current Profiler data
 \item adcp2 - MMS Acoustic Doppler Current Profiler data
 \item cwind - Continuous Winds data
 \item dart - Deep-ocean Assessment and Reporting of Tsunamis data
 \item mmbcur - Marsh-McBirney Current Measurements data
 \item ocean - Oceanographic data
 \item pwind - Peak Winds data
 \item stdmet- Standard Meteorological data
 \item swden - Spectral Wave Density data with Spectral Wave Direction data
 \item wlevel - Water Level data
}
}
\examples{
\dontrun{
buoy(dataset = 'cwind', buoyid = 46085)
buoy(dataset = 'pwind', buoyid = 41021)

#curl debugging
library('httr')
buoy(dataset = 'cwind', buoyid = 46085, config=verbose())
}
}
\references{
\url{http://www.ndbc.noaa.gov/} and \url{http://dods.ndbc.noaa.gov/}
}
\seealso{
\link{buoy_buoys}, \link{buoy_files}, \link{buoy_single_file_url},
\link{get_ncdf_file}, \link{buoy_collect_data}
}

