% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buoy_collect_data}
\alias{buoy_collect_data}
\title{Download a single ncdf file}
\usage{
buoy_collect_data(path)
}
\arguments{
\item{path}{Path to a single buoy data file on local system}
}
\description{
Download a single ncdf file
}
\keyword{internal}

