% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{caching}
\alias{caching}
\alias{ghcnd_clear_cache}
\title{Clear cached files}
\usage{
ghcnd_clear_cache(force = FALSE)
}
\arguments{
\item{force}{(logical) Should we force removal of files if permissions
say otherwise?. Default: \code{FALSE}}
}
\description{
Clear cached files
}
\details{
BEWARE: this will clear all cached files.
}
\section{File storage}{

We use \pkg{rappdirs} to store files, see
\code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}} for how
we determine the directory on your machine to save files to, and run
\code{user_cache_dir("rnoaa")} to get that directory.
}

