% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_response}
\alias{check_response}
\title{Check response from NOAA, including status codes, server error messages, mime-type, etc.}
\usage{
check_response(x)
}
\description{
Check response from NOAA, including status codes, server error messages, mime-type, etc.
}
\keyword{internal}

