% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\name{check_response_erddap}
\alias{check_response_erddap}
\title{Check response from NOAA, including status codes, server error messages, mime-type, etc.}
\usage{
check_response_erddap(x)
}
\description{
Check response from NOAA, including status codes, server error messages, mime-type, etc.
}
\keyword{internal}

