% Generated by roxygen2 (4.0.0): do not edit by hand
\name{check_response_erdddap}
\alias{check_response_erdddap}
\title{Check response from NOAA, including status codes, server error messages, mime-type, etc.}
\usage{
check_response_erdddap(x)
}
\description{
Check response from NOAA, including status codes, server error messages, mime-type, etc.
}
\keyword{internal}

