% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{clean_ghncd_element}
\alias{clean_ghncd_element}
\title{Restructure element of ghncd_search list}
\usage{
clean_ghncd_element(x, keep_flags = FALSE)
}
\arguments{
\item{x}{A dataframe with daily observations for a single monitor for a
single weather variable. This dataframe is one of the elements returned
by \code{\link{ghncd_search}}.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all the flags
for each weather varialbe. The default is to not keep the flags (FALSE).}
}
\value{
A dataframe reformatted to allow easy aggregation of all weather
   variables for a single monitor.
}
\description{
This function restructures a single element of the list object created
by \code{\link{ghncd_search}}, to add a column giving the variable name
(\code{key}) and change the name of the variable column to \code{value}.
These changes facilitate combining all elements from the list created by
\code{\link{ghncd_search}}, to create a tidy dataframe of the weather
observations from the station.
}

