% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coops.R
\name{coops}
\alias{coops}
\alias{coops_search}
\title{Get NOAA co-ops data}
\usage{
coops_search(begin_date = NULL, end_date = NULL, station_name = NULL,
  product = NULL, datum = NULL, units = "metric", time_zone = "gmt",
  ...)
}
\arguments{
\item{begin_date}{(numeric) Date in yyyymmdd format. Required}

\item{end_date}{(numeric) Date in yyyymmdd format. Required}

\item{station_name}{(numeric)}

\item{product}{(character) Specify the data type. See below for Details. Required}

\item{datum}{(character) See below for Details. Required for all water level products.}

\item{units}{(character) Specify metric or english (imperial) units, one of 'metric', 'english'.}

\item{time_zone}{(character) Time zone, one of 'gmt', 'lst', 'lst_ldt'.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}. Optional}
}
\description{
Get NOAA co-ops data
}
\details{
Options for the product paramter. One of:
\itemize{
 \item water_level - Preliminary or verified water levels, depending on availability
 \item air_temperature - Air temperature as measured at the station
 \item water_temperature - Water temperature as measured at the station
 \item wind - Wind speed, direction, and gusts as measured at the station
 \item air_pressure - Barometric pressure as measured at the station
 \item air_gap - Air Gap (distance between a bridge and the water's surface) at the station
 \item conductivity - The water's conductivity as measured at the station
 \item visibility - Visibility from the station's visibility sensor. A measure of atmospheric clarity
 \item humidity - Relative humidity as measured at the station
 \item salinity - Salinity and specific gravity data for the station
 \item hourly_height - Verified hourly height water level data for the station
 \item high_low - Verified high/low water level data for the station
 \item daily_mean - Verified daily mean water level data for the station
 \item monthly_mean - Verified monthly mean water level data for the station
 \item one_minute_water_level - One minute water level data for the station
 \item predictions - 6 minute predictions water level data for the station
 \item datums - datums data for the stations
 \item currents - Currents data for currents stations
}

Options for the datum paramter. One of:
\itemize{
 \item MHHW - Mean higher high water
 \item MHW - Mean high water
 \item MTL - Mean tide level
 \item MSL - Mean sea level
 \item MLW - Mean low water
 \item MLLW - Mean lower low water
 \item NAVD - North American Vertical Datum
 \item STND - Station datum
}
}
\examples{
\dontrun{
coops_search(station_name = 8723970, begin_date = 19820301, end_date = 20141001,
 datum = "stnd", product = "monthly_mean")
}
}
\references{
\url{http://co-ops.nos.noaa.gov/api/}
}

