% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert from degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{A numeric vector in units of degrees.}
}
\value{
The input numeric vector, converted to units of radians.
}
\description{
Convert from degrees to radians
}
