% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disk}
\alias{disk}
\alias{memory}
\title{Options for saving ERDDAP datasets.}
\usage{
disk(path = "~/.rnoaa/erddap", overwrite = TRUE)

memory()
}
\arguments{
\item{path}{Path to store files in. Default: ~/.rnoaa/upwell}

\item{overwrite}{(logical) Overwrite an existing file of the same name? Default: TRUE}
}
\description{
Options for saving ERDDAP datasets.
}

