% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-functions.R
\name{eq_map}
\alias{eq_map}
\title{Displays an R leaflet map with the quake's epicenters (requires internet
access).}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
object with the earthquake epicenters.}

\item{annot_col}{A character vector in HTML format with the text content to
be displayed in the popup text labels of the R leaflet map.}
}
\description{
\code{eq_map} shows an R leaflet map with the quakes' epicenters; it takes
into account the earthquake's magnitude to display the radii.  This function
allows to include popup labels based on the contents of a variable (e.g., the
quake's occurrence).
}
\section{Warning}{

In order to work, the following variables must exist in the \code{data}
object: LONGITUDE, LATITUDE, and EQ_PRIMARY (magnitude).
}

\examples{
require(dplyr)
require(lubridate)

# Before showing the interactive map, we need to tidy the data up.
raw_data <- get_earthquake_data()
clean_data <- eq_clean_data(raw_data)
clean_data <- eq_location_clean(clean_data)

# Displays an R's leaflet map with the epicenters of the earthquakes that
# have occurred in Mexico as of 2000.  The interactive map has popup text
# labels with the date of occurrence.
clean_data \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
}
