% Generated by roxygen2 (4.0.0): do not edit by hand
\name{erddap}
\alias{erddap}
\title{ERDDAP searches and data retrieval}
\usage{
erddap(datasetid, fields = NULL, ..., callopts = list())
}
\arguments{
\item{datasetid}{Dataset id}

\item{fields}{Columns to return, as a character vector}

\item{...}{}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}
}
\description{
ERDDAP searches and data retrieval
}
\examples{
\dontrun{
erddap(datasetid='erdCalCOFIfshsiz', fields=c('longitude','latitude','fish_size','itis_tsn'),
   'time>=' = '2001-07-07','time<=' = '2001-07-10')
erddap(datasetid='erdCinpKfmBT', fields=c('latitude','longitude','Aplysia_californica_Mean_Density','Muricea_californica_Mean_Density'),
   'time>=' = '2007-06-24','time<=' = '2007-07-01')
}
}

