% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caching.R
\name{erddap_clear_cache}
\alias{erddap_clear_cache}
\title{Clear cached files}
\usage{
erddap_clear_cache(path = disk())
}
\arguments{
\item{path}{Path to location of cached files. Defaults to \code{disk()$path}}
}
\description{
Clear cached files
}
\details{
BEWARE: this will clear all cached files.
}

