% Generated by roxygen2 (4.0.0): do not edit by hand
\name{erddap_data}
\alias{erddap_data}
\title{Get ERDDAP data.}
\usage{
erddap_data(datasetid, fields = NULL, ..., callopts = list())
}
\arguments{
\item{datasetid}{Dataset id}

\item{fields}{Columns to return, as a character vector}

\item{...}{Any numbe rof key-value pairs in quotes. See examples}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}
}
\description{
Get ERDDAP data.
}
\examples{
\dontrun{
erddap_data(datasetid='erdCalCOFIfshsiz', fields=c('longitude','latitude','fish_size','itis_tsn'),
   'time>=' = '2001-07-07','time<=' = '2001-07-10')
erddap_data(datasetid='erdCinpKfmBT', fields=c('latitude','longitude',
   'Aplysia_californica_Mean_Density','Muricea_californica_Mean_Density'),
   'time>=' = '2007-06-24','time<=' = '2007-07-01')

# An example workflow
(out <- erddap_search(query='fish size'))
id <- out$info$dataset_id[1]
erddap_info(datasetid=id)$variables
erddap_data(datasetid = id, fields = c('latitude','longitude','scientific_name'),
   'time>=' = '2001-07-14')
}
}

