% Generated by roxygen2 (4.0.0): do not edit by hand
\name{erddap_info}
\alias{erddap_info}
\title{Get information on an ERDDAP dataset.}
\usage{
erddap_info(datasetid, callopts = list())
}
\arguments{
\item{datasetid}{Dataset id}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}
}
\description{
Get information on an ERDDAP dataset.
}
\examples{
\dontrun{
erddap_info(datasetid='erdCalCOFIfshsiz')
out <- erddap_info(datasetid='erdCinpKfmBT')
}
}

