% Generated by roxygen2 (4.0.2): do not edit by hand
\name{erddap_search}
\alias{erddap_search}
\title{Search for ERDDAP datasets.}
\usage{
erddap_search(query, page = NULL, page_size = NULL, callopts = list())
}
\arguments{
\item{query}{Search terms}

\item{page}{Page number}

\item{page_size}{Results per page}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}
}
\description{
Search for ERDDAP datasets.
}
\examples{
\dontrun{
(out <- erddap_search(query='fish size'))
out$alldata
(out <- erddap_search(query='size'))
out$info
}
}

