% Generated by roxygen2 (4.0.2): do not edit by hand
\name{erddap_upwell_info}
\alias{erddap_upwell_info}
\title{Get information on an ERDDAP dataset.}
\usage{
erddap_upwell_info(datasetid, ...)
}
\arguments{
\item{datasetid}{Dataset id}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}
}
\value{
A list of length two
\itemize{
 \item variables Data.frame of variables and their types
 \item alldata List of data variables and their full attributes
}
}
\description{
Gives back a brief data.frame for quick inspection and a list of more complete
metadata on the dataset.
}
\examples{
\dontrun{
erddap_upwell_info('noaa_pfeg_696e_ec99_6fa6')
erddap_upwell_info(datasetid='noaa_ngdc_34bf_a95c_7e28')

(out <- erddap_upwell_search(query='temperature'))
erddap_upwell_info(out$info$dataset_id[5])
erddap_upwell_info(out$info$dataset_id[15])
erddap_upwell_info(out$info$dataset_id[25])
erddap_upwell_info(out$info$dataset_id[33])
erddap_upwell_info(out$info$dataset_id[65])
erddap_upwell_info(out$info$dataset_id[150])
erddap_upwell_info(out$info$dataset_id[400])
erddap_upwell_info(out$info$dataset_id[678])

out <- erddap_upwell_info(datasetid='noaa_ngdc_34bf_a95c_7e28')
## See brief overview of the variables and range of possible values, if given
out$variables
## all information on longitude
out$alldata$longitude
## all information on Climatological_Temperature
out$alldata$Climatological_Temperature
}
}

