% Generated by roxygen2 (4.0.2): do not edit by hand
\name{erddap_upwell_search}
\alias{erddap_upwell_search}
\title{Search for ERDDAP UPWELL datasets.}
\usage{
erddap_upwell_search(query, page = NULL, page_size = NULL, ...)
}
\arguments{
\item{query}{Search terms}

\item{page}{Page number}

\item{page_size}{Results per page}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must be a named parameter)}
}
\description{
Search for ERDDAP UPWELL datasets.
}
\examples{
\dontrun{
(out <- erddap_upwell_search(query='temperature'))
out$alldata[[1]]
(out <- erddap_upwell_search(query='size'))
out$info
}
}

