% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnoaa-package.r
\docType{data}
\name{fipscodes}
\alias{fipscodes}
\title{FIPS codes for US states.}
\format{A data frame with 3142 rows and 5 variables}
\description{
A dataset containing the FIPS codes for 51 US states
		and territories. The variables are as follows:
}
\details{
\itemize{
  \item state. US state name.
  \item county. County name.
  \item fips_state. Numeric value, from 1 to 51.
  \item fips_county. Numeric value, from 1 to 840.
  \item fips. Numeric value, from 1001 to 56045.
}
}
\keyword{datasets}
