% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timelinelabel.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Class of the geom_timeline_label}
\description{
Please refer to the documentation of the \code{\link{geom_timeline_label}}.
}
\keyword{datasets}
