% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{get_earthquake_data}
\alias{get_earthquake_data}
\title{Loads the NOAA's Significant Earthquake dataset.}
\usage{
get_earthquake_data()
}
\value{
Returns a \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
object.
}
\description{
It reads the \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{Significant Earthquake database}
from a plain file downloaded on September 10, 2017 from the NOAA's Webpage
(i.e., no internet connection is needed); in case you want the latest
information, you can use the \code{\link{download_earthquake_data}} function.
}
\details{
This dataset contains information about destructive earthquakes from 2150
B.C. to the present that meet the following criterion: Moderate damage
(approximately $1 million or more), 10 or more deaths, Magnitude 7.5 or
greater, Modified Mercalli Intensity X or greater, or the earthquake
generated a tsunami.

Please refer to \href{https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225}{this link}
for more information about the description of each variable.
}
\examples{
raw_data <- get_earthquake_data()
}
\seealso{
\code{\link{eq_clean_data}} and eq_location_clean for data tidying.
}
