% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_ncdf_file}
\alias{get_ncdf_file}
\title{Download a single ncdf file}
\usage{
get_ncdf_file(path, buoyid, file, output)
}
\arguments{
\item{path}{Path to a single buoy data file}

\item{buoyid}{Buoy id. (optional)}

\item{file}{XX}

\item{output}{XX}
}
\description{
Download a single ncdf file
}
\keyword{internal}

