% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ghcnd.R
\name{ghcnd}
\alias{ghcnd}
\title{Get GHCND daily data from NOAA FTP server}
\usage{
ghcnd(stationid, path = "~/.rnoaa/ghcnd", overwrite = TRUE, ...)
}
\arguments{
\item{stationid}{Stationid to get}

\item{path}{(character) A path to store the files, Default: \code{~/.rnoaa/isd}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get GHCND daily data from NOAA FTP server
}
\examples{
\dontrun{
# Get stations
stations <- ghcnd_stations()
head(stations)

# Get data
ghcnd(stationid="AGE00147704")
ghcnd(stations$id[40])
ghcnd(stations$id[4000])
ghcnd(stations$id[10000])
ghcnd(stations$id[80000])

library("dplyr")
dat <- ghcnd(stations$id[10000])
dat$data \%>\%
   filter(element == "PRCP", year == 1884)
}
}

