% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{is.ncdc_data}
\alias{is.ncdc_data}
\alias{is.ncdc_datasets}
\alias{is.ncdc_datatypes}
\alias{is.ncdc_datacats}
\alias{is.ncdc_locs}
\alias{is.ncdc_locs_cats}
\alias{is.ncdc_stations}
\title{Check object class}
\usage{
is.ncdc_data(x)

is.ncdc_datasets(x)

is.ncdc_datatypes(x)

is.ncdc_datacats(x)

is.ncdc_locs(x)

is.ncdc_locs_cats(x)

is.ncdc_stations(x)
}
\arguments{
\item{x}{input}
}
\description{
Check if an object is of class ncdc_data, ncdc_datasets,
ncdc_datatypes, ncdc_datacats, ncdc_locs, ncdc_locs_cats,
or ncdc_stations
}
\keyword{internal}
