\name{is.noaa_data}
\alias{is.noaa_data}
\alias{is.noaa_datacats}
\alias{is.noaa_datasets}
\alias{is.noaa_datatypes}
\alias{is.noaa_locs}
\alias{is.noaa_locs_cats}
\alias{is.noaa_stations}
\title{Check object class}
\usage{
is.noaa_data(x)

is.noaa_datasets(x)

is.noaa_datatypes(x)

is.noaa_datacats(x)

is.noaa_locs(x)

is.noaa_locs_cats(x)

is.noaa_stations(x)
}
\arguments{
  \item{x}{input}
}
\description{
Check if an object is of class noaa_data, noaa_datasets,
noaa_datatypes, noaa_datacats, noaa_locs, noaa_locs_cats,
or noaa_stations
}

