% Generated by roxygen2 (4.0.0): do not edit by hand
\name{latlong2bbox}
\alias{latlong2bbox}
\title{Function to calculate bounding box for the extent parameter in noaa_stations function.}
\usage{
latlong2bbox(lat, lon, radius = 10)
}
\description{
Function to calculate bounding box for the extent parameter in noaa_stations function.
}
\examples{
latlong2bbox(lat=33.95, lon=-118.40) # radius of 10 km
latlong2bbox(lat=33.95, lon=-118.40, radius=2) # radius of 2 km
latlong2bbox(lat=33.95, lon=-118.40, radius=200) # radius of 200 km
latlong2bbox(lat=33.95, lon=-118.40, radius=0.02) # radius of 20 meters
}
\keyword{internal}

