% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd.R
\name{lcd}
\alias{lcd}
\title{Local Climitalogical Data from NOAA}
\usage{
lcd(station, year, ...)
}
\arguments{
\item{station}{(character) station code, e.g., "02413099999". we will
allow integer/numeric passed here, but station ids can have leading
zeros, so it's a good idea to keep stations as character class}

\item{year}{(integer) year, e.g., 2017}

\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a data.frame, with many columns, and variable rows
depending on how frequently data was collected in the given year
}
\description{
Local Climitalogical Data from NOAA
}
\examples{
\dontrun{
lcd(station = "01338099999", year = "2017")
lcd(station = "01338099999", year = "2015")

lcd(station = "02413099999", year = "2009")
lcd(station = "02413099999", year = "2001")

# pass curl options
lcd(station = "02413099999", year = "2002", verbose = TRUE)
}
}
\references{
\url{https://www.ncdc.noaa.gov/cdo-web/datatools/lcd}
\url{https://www1.ncdc.noaa.gov/pub/data/cdo/documentation/LCD_documentation.pdf}
}
