% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd.R
\docType{data}
\name{lcd_cache}
\alias{lcd_cache}
\title{lcd_cache}
\format{
An object of class \code{HoardClient} (inherits from \code{R6}) of length 19.
}
\usage{
lcd_cache
}
\description{
Manage the \code{lcd()} cache
}
\details{
The cache directory for \code{lcd()} is
\code{paste0(rappdirs::user_cache_dir(), "/R/noaa_lcd")}, but you can set
your own path using \code{cache_path_set()}

\code{cache_delete} only accepts 1 file name, while
\code{cache_delete_all} doesn't accept any names, but deletes all files.
For deleting many specific files, use \code{cache_delete} in a \code{\link[=lapply]{lapply()}}
type call
}
\section{Useful user functions}{

\itemize{
\item \code{lcd_cache$cache_path_get()} get cache path
\item \code{lcd_cache$cache_path_set()} set cache path
\item \code{lcd_cache$list()} returns a character vector of full path file names
\item \code{lcd_cache$files()} returns file objects with metadata
\item \code{lcd_cache$details()} returns files with details
\item \code{lcd_cache$delete()} delete specific files
\item \code{lcd_cache$delete_all()} delete all files, returns nothing
}
}

\examples{
\dontrun{
lcd_cache

# list files in cache
lcd_cache$list()

# delete certain database files
# lcd_cache$delete("file path")
# lcd_cache$list()

# delete all files in cache
# lcd_cache$delete_all()
# lcd_cache$list()

# set a different cache path from the default
# lcd_cache$cache_path_set(full_path = file.path(tempdir(), "foo_bar"))
# lcd_cache
}
}
\keyword{datasets}
