% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd.R
\name{lcd_columns}
\alias{lcd_columns}
\title{Specify Variable Types in Local Climatological Data from NOAA}
\usage{
lcd_columns(
  STATION = "character",
  DATE = "POSIXct",
  LATITUDE = "numeric",
  LONGITUDE = "numeric",
  ELEVATION = "numeric",
  NAME = "character",
  REPORT_TYPE = "character",
  SOURCE = "character",
  HourlyAltimeterSetting = "character",
  HourlyDewPointTemperature = "character",
  HourlyDryBulbTemperature = "character",
  HourlyPrecipitation = "character",
  HourlyPresentWeatherType = "character",
  HourlyPressureChange = "character",
  HourlyPressureTendency = "integer",
  HourlyRelativeHumidity = "character",
  HourlySkyConditions = "character",
  HourlySeaLevelPressure = "character",
  HourlyStationPressure = "character",
  HourlyVisibility = "character",
  HourlyWetBulbTemperature = "character",
  HourlyWindDirection = "character",
  HourlyWindGustSpeed = "character",
  HourlyWindSpeed = "character",
  Sunrise = "numeric",
  Sunset = "numeric",
  DailyAverageDewPointTemperature = "character",
  DailyAverageDryBulbTemperature = "character",
  DailyAverageRelativeHumidity = "character",
  DailyAverageSeaLevelPressure = "character",
  DailyAverageStationPressure = "character",
  DailyAverageWetBulbTemperature = "character",
  DailyAverageWindSpeed = "character",
  DailyCoolingDegreeDays = "numeric",
  DailyDepartureFromNormalAverageTemperature = "numeric",
  DailyHeatingDegreeDays = "numeric",
  DailyMaximumDryBulbTemperature = "numeric",
  DailyMinimumDryBulbTemperature = "numeric",
  DailyPeakWindDirection = "numeric",
  DailyPeakWindSpeed = "numeric",
  DailyPrecipitation = "character",
  DailySnowDepth = "character",
  DailySnowfall = "character",
  DailySustainedWindDirection = "numeric",
  DailySustainedWindSpeed = "numeric",
  DailyWeather = "character",
  MonthlyAverageRH = "numeric",
  MonthlyDaysWithGT001Precip = "numeric",
  MonthlyDaysWithGT010Precip = "numeric",
  MonthlyDaysWithGT32Temp = "numeric",
  MonthlyDaysWithGT90Temp = "numeric",
  MonthlyDaysWithLT0Temp = "numeric",
  MonthlyDaysWithLT32Temp = "numeric",
  MonthlyDepartureFromNormalAverageTemperature = "numeric",
  MonthlyDepartureFromNormalCoolingDegreeDays = "numeric",
  MonthlyDepartureFromNormalHeatingDegreeDays = "numeric",
  MonthlyDepartureFromNormalMaximumTemperature = "numeric",
  MonthlyDepartureFromNormalMinimumTemperature = "numeric",
  MonthlyDepartureFromNormalPrecipitation = "numeric",
  MonthlyDewpointTemperature = "numeric",
  MonthlyGreatestPrecip = "numeric",
  MonthlyGreatestPrecipDate = "character",
  MonthlyGreatestSnowDepth = "numeric",
  MonthlyGreatestSnowDepthDate = "character",
  MonthlyGreatestSnowfall = "numeric",
  MonthlyGreatestSnowfallDate = "character",
  MonthlyMaxSeaLevelPressureValue = "numeric",
  MonthlyMaxSeaLevelPressureValueDate = "character",
  MonthlyMaxSeaLevelPressureValueTime = "character",
  MonthlyMaximumTemperature = "numeric",
  MonthlyMeanTemperature = "numeric",
  MonthlyMinSeaLevelPressureValue = "numeric",
  MonthlyMinSeaLevelPressureValueDate = "character",
  MonthlyMinSeaLevelPressureValueTime = "character",
  MonthlyMinimumTemperature = "numeric",
  MonthlySeaLevelPressure = "numeric",
  MonthlyStationPressure = "numeric",
  MonthlyTotalLiquidPrecipitation = "numeric",
  MonthlyTotalSnowfall = "numeric",
  MonthlyWetBulb = "numeric",
  AWND = "numeric",
  CDSD = "numeric",
  CLDD = "numeric",
  DSNW = "numeric",
  HDSD = "numeric",
  HTDD = "numeric",
  NormalsCoolingDegreeDay = "numeric",
  NormalsHeatingDegreeDay = "numeric",
  ShortDurationEndDate005 = "character",
  ShortDurationEndDate010 = "character",
  ShortDurationEndDate015 = "character",
  ShortDurationEndDate020 = "character",
  ShortDurationEndDate030 = "character",
  ShortDurationEndDate045 = "character",
  ShortDurationEndDate060 = "character",
  ShortDurationEndDate080 = "character",
  ShortDurationEndDate100 = "character",
  ShortDurationEndDate120 = "character",
  ShortDurationEndDate150 = "character",
  ShortDurationEndDate180 = "character",
  ShortDurationPrecipitationValue005 = "numeric",
  ShortDurationPrecipitationValue010 = "numeric",
  ShortDurationPrecipitationValue015 = "numeric",
  ShortDurationPrecipitationValue020 = "numeric",
  ShortDurationPrecipitationValue030 = "numeric",
  ShortDurationPrecipitationValue045 = "numeric",
  ShortDurationPrecipitationValue060 = "numeric",
  ShortDurationPrecipitationValue080 = "numeric",
  ShortDurationPrecipitationValue100 = "numeric",
  ShortDurationPrecipitationValue120 = "numeric",
  ShortDurationPrecipitationValue150 = "numeric",
  ShortDurationPrecipitationValue180 = "numeric",
  REM = "character",
  BackupDirection = "character",
  BackupDistance = "character",
  BackupDistanceUnit = "character",
  BackupElements = "character",
  BackupElevation = "character",
  BackupEquipment = "character",
  BackupLatitude = "character",
  BackupLongitude = "character",
  BackupName = "character",
  WindEquipmentChangeDate = "character"
)
}
\arguments{
\item{STATION}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{LATITUDE}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{LONGITUDE}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ELEVATION}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{NAME}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{REPORT_TYPE}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{SOURCE}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyAltimeterSetting}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyDewPointTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyDryBulbTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyPrecipitation}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyPresentWeatherType}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyPressureChange}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyPressureTendency}{(character) string indicating variable or column type that is returned, default is "integer". optional}

\item{HourlyRelativeHumidity}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlySkyConditions}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlySeaLevelPressure}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyStationPressure}{(character)string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyVisibility}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyWetBulbTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyWindDirection}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyWindGustSpeed}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{HourlyWindSpeed}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{Sunrise}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{Sunset}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyAverageDewPointTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageDryBulbTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageRelativeHumidity}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageSeaLevelPressure}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageStationPressure}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageWetBulbTemperature}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyAverageWindSpeed}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailyCoolingDegreeDays}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyDepartureFromNormalAverageTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyHeatingDegreeDays}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyMaximumDryBulbTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyMinimumDryBulbTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyPeakWindDirection}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyPeakWindSpeed}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyPrecipitation}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailySnowDepth}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailySnowfall}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{DailySustainedWindDirection}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailySustainedWindSpeed}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DailyWeather}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyAverageRH}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithGT001Precip}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithGT010Precip}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithGT32Temp}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithGT90Temp}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithLT0Temp}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDaysWithLT32Temp}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalAverageTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalCoolingDegreeDays}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalHeatingDegreeDays}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalMaximumTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalMinimumTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDepartureFromNormalPrecipitation}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyDewpointTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyGreatestPrecip}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyGreatestPrecipDate}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyGreatestSnowDepth}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyGreatestSnowDepthDate}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyGreatestSnowfall}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyGreatestSnowfallDate}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyMaxSeaLevelPressureValue}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyMaxSeaLevelPressureValueDate}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyMaxSeaLevelPressureValueTime}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyMaximumTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyMeanTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyMinSeaLevelPressureValue}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyMinSeaLevelPressureValueDate}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyMinSeaLevelPressureValueTime}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{MonthlyMinimumTemperature}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlySeaLevelPressure}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyStationPressure}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyTotalLiquidPrecipitation}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyTotalSnowfall}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{MonthlyWetBulb}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{AWND}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{CDSD}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{CLDD}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{DSNW}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{HDSD}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{HTDD}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{NormalsCoolingDegreeDay}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{NormalsHeatingDegreeDay}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationEndDate005}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate010}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate015}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate020}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate030}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate045}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate060}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate080}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate100}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate120}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate150}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationEndDate180}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{ShortDurationPrecipitationValue005}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue010}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue015}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue020}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue030}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue045}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue060}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue080}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue100}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue120}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue150}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{ShortDurationPrecipitationValue180}{(character) string indicating variable or column type that is returned, default is "numeric". optional}

\item{REM}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupDirection}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupDistance}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupDistanceUnit}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupElements}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupElevation}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupEquipment}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupLatitude}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupLongitude}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{BackupName}{(character) string indicating variable or column type that is returned, default is "character". optional}

\item{WindEquipmentChangeDate}{(character) string indicating variable or column type that is returned, default is "character". optional}
}
\value{
a vector indicating column classes types
}
\description{
Use this function to specify what variable types will be returned by
\link{lcd}. The function returns a named vector with specified column
classes. The defaults are specified in the argument descriptions below.
}
\note{
if the column type is not compatible, \link{lcd} will return a
dataframe with the most appropriate column type and a message indicating
the column was not changed to the specified type.
}
\keyword{internal}
