% Generated by roxygen2 (4.0.0): do not edit by hand
\name{long2utm}
\alias{long2utm}
\title{Function to get UTM zone from a single longitude and latitude pair
originally from David LeBauer I think}
\usage{
long2utm(lon, lat)
}
\description{
Function to get UTM zone from a single longitude and latitude pair
originally from David LeBauer I think
}

