% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_cache.r
\name{meteo_clear_cache}
\alias{meteo_clear_cache}
\title{Clear \emph{meteo} cached files}
\usage{
meteo_clear_cache(force = FALSE)
}
\arguments{
\item{force}{(logical) force delete. default: \code{FALSE}}
}
\description{
The \emph{meteo} functions use an aplication
}
\note{
This function will clear all cached \emph{meteo} files.
}
\seealso{
Other meteo: \code{\link{meteo_show_cache}}
}
