% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_utils.r
\name{meteo_coverage}
\alias{meteo_coverage}
\title{Determine the "coverage" for a station data frame}
\usage{
meteo_coverage(meteo_df, obs_start_date = NULL, obs_end_date = NULL,
  verbose = TRUE)
}
\arguments{
\item{meteo_df}{an \emph{meteo} \code{data.frame}}

\item{obs_end_date, obs_end_date}{specify either or both to constrain
coverate tests. These should be \code{Date} objects.}

\item{verbose}{if \code{TRUE} will display the coverage summary along
with returning the coverage data.frame}
}
\value{
a \code{data.frame} with the coverage for each station, minimally
containing: \preformatted{
$ id         (fctr)
$ station    (fctr)
$ start_date (time)
$ end_date   (time)
$ total_obs  (int)
}
}
\description{
Call this function after pulling down observations for a set of stations
to retrieve the "coverage" (i.e. how complete each field is). If either
or both \code{obs_start_date} or \code{obs_end_date} are specified,
the coverage test will be limited to that date range.
}

