% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_distance}
\alias{meteo_distance}
\title{meteo_distance}
\usage{
meteo_distance(station_data, lat, long, latName = "latitude",
  longName = "longitude", units = "deg", radius = NULL, limit = NULL)
}
\arguments{
\item{station_data}{a dataframe. Expects col headers with names latName and longName}

\item{lat}{Latitude to centre search at}

\item{long}{Longitude to centre search at}

\item{latName}{Name of latitude header name in data, Default = 'latitude'}

\item{longName}{Name of longitude header name in data. Default = 'longitude'}

\item{units}{Units of the latitude and longitude values: degrees 'deg',
radians 'rad', d/m/s 'dms'. Default = 'deg'}

\item{radius}{Radius to search (does nothing yet)}

\item{limit}{Upperbound on number of results. Deafult = 1}
}
\value{
a dataframe in a column with the distance of stations
}
\description{
meteo_distance
}

