% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{meteo_tidy_ghcnd_element}
\alias{meteo_tidy_ghcnd_element}
\title{Restructure element of ghcnd_search list}
\usage{
meteo_tidy_ghcnd_element(x, keep_flags = FALSE)
}
\arguments{
\item{x}{A dataframe with daily observations for a single monitor for a
single weather variable. This dataframe is one of the elements returned
by \code{\link{ghcnd_search}}.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all
the flags for each weather variable. The default is to not keep the
flags (FALSE). See the note below for more information on these flags.}
}
\value{
A dataframe reformatted to allow easy aggregation of all weather
   variables for a single monitor.
}
\description{
This function restructures the output of \code{\link{ghcnd_search}}
to add a column giving the variable name (\code{key}) and change the
name of the variable column to \code{value}. These changes facilitate
combining all elements from the list created by \code{\link{ghcnd_search}},
to create a tidy dataframe of the weather observations from the station.
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu}
}
