% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexrad.R
\name{nexrad2}
\alias{nexrad2}
\title{NEXRAD II data}
\usage{
nexrad2(year, month, day, station, filename, out, ...)
}
\arguments{
\item{x}{xx}
}
\value{
data.frame
}
\description{
NEXRAD II data
}
\details{
xx
}
\examples{
\dontrun{
f <- tempfile(fileext = ".zip")
nexrad2(2017, 4, 1, 'KAMX', 'KAMX20170401_002500_V06', f)
}
}
