% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaa_buoy_buoys}
\alias{noaa_buoy_buoys}
\title{Get NOAA buoy data from the National Buoy Data Center}
\usage{
noaa_buoy_buoys(dataset = NULL, ...)
}
\arguments{
\item{dataset}{Dataset to query. See below for details. (required)}

\item{...}{Further arguments passed on to the API GET call. (optional)}
}
\description{
Get NOAA buoy data from the National Buoy Data Center
}
\keyword{internal}

