% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaa_buoy_files}
\alias{noaa_buoy_files}
\title{Get NOAA buoy data from the National Buoy Data Center}
\usage{
noaa_buoy_files(path, buoyid, ...)
}
\arguments{
\item{path}{Path to a single buoy data file}

\item{buoyid}{Buoy id. (optional)}

\item{...}{Further arguments passed on to the API GET call. (optional)}
}
\description{
Get NOAA buoy data from the National Buoy Data Center
}
\keyword{internal}

