% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaa_buoy_single_file_url}
\alias{noaa_buoy_single_file_url}
\title{Make url for a single NOAA buoy data file}
\usage{
noaa_buoy_single_file_url(dataset, buoyid, file)
}
\arguments{
\item{dataset}{Dataset to query. See below for details. (required)}

\item{buoyid}{Buoy id. (optional)}

\item{file}{Output file name}
}
\description{
Make url for a single NOAA buoy data file
}
\keyword{internal}

