% Generated by roxygen2 (4.0.0): do not edit by hand
\name{noaa_datasets}
\alias{noaa_datasets}
\title{Search NOAA datasets}
\usage{
noaa_datasets(datasetid = NULL, datatypeid = NULL, stationid = NULL,
  locationid = NULL, startdate = NULL, enddate = NULL, sortfield = NULL,
  sortorder = NULL, limit = 25, offset = NULL, callopts = list(),
  token = NULL, dataset = NULL, page = NULL, year = NULL,
  month = NULL)
}
\arguments{
  \item{datasetid}{(optional) Accepts a single valid
  dataset id. Data returned will be from the dataset
  specified, see datasets()}

  \item{startdate}{(optional) Accepts valid ISO formated
  date (yyyy-mm-dd). Data returned will have data after the
  specified date. Paramater can be use independently of
  enddate}

  \item{enddate}{(optional) Accepts valid ISO formated date
  (yyyy-mm-dd). Data returned will have data before the
  specified date. Paramater can be use independently of
  startdate}

  \item{dataset}{THIS IS A DEPRECATED ARGUMENT. See
  datasetid.}

  \item{page}{THIS IS A DEPRECATED ARGUMENT. There is no
  equivalent argument in v2 of the NOAA API.}

  \item{year}{THIS IS A DEPRECATED ARGUMENT. Use
  combination of startdate and enddate arguments.}

  \item{month}{THIS IS A DEPRECATED ARGUMENT. Use
  combination of startdate and enddate arguments.}
}
\value{
A data.frame for all datasets, or a list of length two, each with a data.frame.
}
\description{
From the NOAA API docs: All of our data are in datasets. To retrieve any data
from us, you must know what dataset it is in.
}
\examples{
\dontrun{
# Get a table of all datasets
noaa_datasets()

# Get details from a particular dataset
noaa_datasets(datasetid='ANNUAL')

# Get datasets with Temperature at the time of observation (TOBS) data type
noaa_datasets(datatypeid='TOBS')

# Get datasets with data for a series of the same parameter arg, in this case
# stationid's
noaa_datasets(stationid=c('COOP:310090','COOP:310184','COOP:310212'))

# Multiple datatypeid's
noaa_datasets(datatypeid=c('ACMC','ACMH','ACSC'))
noaa_datasets(datasetid='ANNUAL', datatypeid=c('ACMC','ACMH','ACSC'))
}
}

