% Generated by roxygen2 (4.0.0): do not edit by hand
\name{noaa_datatypes}
\alias{noaa_datatypes}
\title{Get possible data types for a particular dataset}
\usage{
noaa_datatypes(datasetid = NULL, datatypeid = NULL, datacategoryid = NULL,
  stationid = NULL, locationid = NULL, startdate = NULL, enddate = NULL,
  sortfield = NULL, sortorder = NULL, limit = 25, offset = NULL,
  callopts = list(), token = getOption("noaakey",
  stop("you need an API key NOAA data")), dataset = NULL, page = NULL,
  filter = NULL)
}
\arguments{
  \item{datasetid}{(optional) Accepts a single valid
  dataset id. Data returned will be from the dataset
  specified, see datasets()}

  \item{startdate}{(optional) Accepts valid ISO formated
  date (yyyy-mm-dd). Data returned will have data after the
  specified date. Paramater can be use independently of
  enddate}

  \item{enddate}{(optional) Accepts valid ISO formated date
  (yyyy-mm-dd). Data returned will have data before the
  specified date. Paramater can be use independently of
  startdate}

  \item{dataset}{THIS IS A DEPRECATED ARGUMENT. See
  datasetid.}

  \item{page}{THIS IS A DEPRECATED ARGUMENT. There is no
  equivalent argument in v2 of the NOAA API.}

  \item{filter}{THIS IS A DEPRECATED ARGUMENT. There is no
  equivalent argument in v2 of the NOAA API.}

  \item{datacategoryid}{Optional. Accepts a valid data
  category id or a chain of data category ids seperated by
  ampersands (although it is rare to have a data type with
  more than one data category). Data types returned will be
  associated with the data category(ies) specified}
}
\value{
A \code{data.frame} for all datasets, or a list of length two, each with
   a data.frame.
}
\description{
From the NOAA API docs: Describes the type of data, acts as a label. If it's 64
degrees out right now, then the data type is Air Temperature and the data is 64.
}
\examples{
\dontrun{
# Fetch available data types
noaa_datatypes()

# Fetch more information about the ACMH data type id
noaa_datatypes(datatypeid="ACMH")

# Fetch data types with the air temperature data category
noaa_datatypes(datacategoryid="TEMP", limit=56)

# Fetch data types that support a given set of stations
noaa_datatypes(stationid=c('COOP:310090','COOP:310184','COOP:310212'))
}
}

