% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnd.R
\name{print.ghcnd_stations}
\alias{print.ghcnd_stations}
\title{Print out GHCND stations}
\usage{
\method{print}{ghcnd_stations}(x, ..., n = 10)
}
\arguments{
\item{x}{The object returned by a call to \code{\link{ghcnd_stations}}.}

\item{...}{Other parameters to pass to \code{print}.}

\item{n}{Number of rows of station data to print.}
}
\description{
This function prints out, in a nice format, the object returned by
\code{\link{ghcnd_stations}}.
}
\examples{
\dontrun{
stations <- ghcnd_stations()
print(stations)
}

}

