% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-functions.R
\docType{data}
\name{providers}
\alias{providers}
\title{List of available R's leaflet map providers.}
\description{
Please refer to the documentation of \code{\link[leaflet]{providers}}.
}
\section{Details}{

This object was exported from the leaflet package so as to use different
types of maps in the \code{\link{eq_map}} function.
}

\keyword{datasets}
