% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{rnoaa-defunct}
\alias{rnoaa-defunct}
\title{Defunct functions in rnoaa}
\description{
\itemize{
 \item \code{\link{noaa}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datacats}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datasets}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datatypes}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_locs}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_locs_cats}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_stations}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_plot}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_combine}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_seaice}}: Function name changed to seaice
 \item \code{\link{erddap_data}}: See package rerddap
 \item \code{\link{erddap_clear_cache}}: See package rerddap
 \item \code{\link{erddap_datasets}}: Moved to \code{ed_datasets} in
 package rerddap
 \item \code{\link{erddap_grid}}: Moved to \code{griddap} in package
 rerddap
 \item \code{\link{erddap_info}}: Moved to \code{info} in package
 rerddap
 \item \code{\link{erddap_search}}: Moved to \code{ed_search} in
 package rerddap
 \item \code{\link{erddap_table}}: Moved to \code{tabledap} in
 package rerddap
}
}

