% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnoaa_options.R
\name{rnoaa_options}
\alias{rnoaa_options}
\title{rnoaa options}
\usage{
rnoaa_options(cache_messages = TRUE)
}
\arguments{
\item{cache_messages}{(logical) whether to emit messages with information
on caching status for function calls that can cache data. default: \code{TRUE}}
}
\description{
rnoaa options
}
\details{
rnoaa package level options; stored in an internal
package environment \code{roenv}
}
\examples{
\dontrun{
rnoaa_options(cache_messages = FALSE)
}
}
\seealso{
\link{rnoaa_caching} for managing cached files
}
