% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run a built-in Shiny App.}
\usage{
runShinyApp()
}
\value{
Nothing, it does something.
}
\description{
Run a built-in Shiny App.
}
\examples{
\dontrun{
runShinyApp()
}
}

