% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice.r
\name{sea_ice}
\alias{sea_ice}
\title{Get sea ice data.}
\usage{
sea_ice(year = NULL, month = NULL, pole = NULL, format = "shp",
  ...)
}
\arguments{
\item{year}{(numeric) a year}

\item{month}{(character) a month, as character abbrevation of a month}

\item{pole}{(character) one of S (south) or N (north)}

\item{format}{(character) one of shp (default), geotiff-extent (for geotiff
extent data), or geotiff-conc (for geotiff concentration data)}

\item{...}{Further arguments passed on to \code{rgdal::readshpfile()} if
\code{format="shp"} or \code{raster::raster()} if not}
}
\value{
data.frame if \code{format="shp"}; \code{raster::raster()} if not
}
\description{
Get sea ice data.
}
\details{
For shp files, if you want to reproject the shape files, use
\code{\link[=readshpfile]{readshpfile()}} to read in shape file, then reproject, and so on.
}
\examples{
\dontrun{
# the new way
library(raster)

## one file
sea_ice(year = 1990, month = "Apr", pole = "N")
sea_ice(year = 1990, month = "Apr", pole = "N", format = "geotiff-extent")
sea_ice(year = 1990, month = "Apr", pole = "N", format = "geotiff-conc")
## many files
sea_ice(year = 1990, month = "Apr")
x <- sea_ice(year = 1990, month = "Apr", format = "geotiff-extent")
y <- sea_ice(year = 1990, month = "Apr", format = "geotiff-conc")
plot(x[[1]])
plot(y[[1]])

# Map a single year/month/pole combo
out <- sea_ice(year = 1990, month = 'Apr', pole = 'N')
library('sf')
plot(out[[1]])
}
}
\references{
See the "User Guide" pdf at https://nsidc.org/data/g02135
}
\seealso{
\code{\link[=seaice_tabular]{seaice_tabular()}}
}
