% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice_tabular.R
\name{seaice_tabular}
\alias{seaice_tabular}
\title{Sea ice tabular data}
\usage{
seaice_tabular(...)
}
\arguments{
\item{...}{Curl options passed on to \link[crul:verb-GET]{crul::verb-GET} - beware
that curl options are passed to each http request, for each of
24 requests.}
}
\value{
A data.frame with columns:
\itemize{
\item year (integer)
\item mo (integer)
\item data.type (character)
\item region (character)
\item extent (numeric)
\item area (numeric)
}
}
\description{
Sea ice tabular data
}
\details{
a value in any cell of -9999 indicates missing data
}
\examples{
\dontrun{
df <- seaice_tabular()
df
}
}
\seealso{
\code{\link[=seaice]{seaice()}}
}
