% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{storm_columns}
\alias{storm_columns}
\title{NOAA storm data from International Best Track Archive for Climate Stewardship (IBTrACS).}
\format{A data frame with 195 rows and 8 variables}
\description{
This dataset includes description of the columns of each dataset acquired using
\code{\link[rnoaa]{storm_data}}
}
\keyword{datasets}

