% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.R
\name{storm_data}
\alias{storm_data}
\title{Get NOAA wind storm tabular data, metadata, or shp files from IBTrACS}
\usage{
storm_data(...)
}
\description{
Get NOAA wind storm tabular data, metadata, or shp files from IBTrACS
}
\keyword{internal}
