% Generated by roxygen2 (4.0.2): do not edit by hand
\name{storm_data}
\alias{storm_data}
\title{Get NOAA wind storm data from International Best Track Archive for Climate Stewardship (IBTrACS)}
\usage{
storm_data(basin = NULL, storm = NULL, year = NULL,
  path = "~/.rnoaa/storms", overwrite = TRUE, read = TRUE)
}
\arguments{
\item{basin}{(character) A basin name, one of EP, NA, NI, SA, SI, SP, or WP.}

\item{storm}{(character) A storm serial number of the form YYYYJJJHTTNNN. See Details.}

\item{year}{(numeric) One of the years from 1842 to 2014}

\item{path}{(character) A path to store the files, Default: \code{~/.rnoaa/storms}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}
}
\description{
Get NOAA wind storm data from International Best Track Archive for Climate Stewardship (IBTrACS)
}
\details{
Details for storm serial numbers:
\itemize{
 \item YYYY is the corresponding year of the first recorded observation of the storm
 \item JJJ is the day of year of the first recorded observation of the storm
 \item H is the hemisphere of the storm: N=Northern, S=Southern
 \item TT is the absolute value of the rounded latitude of the first recorded observation of the
 storm (range 0-90, if basin=SA or SH, then TT in reality is negative)
 \item NNN is the rounded longitude of the first recorded observation of the storm (range 0-359)
}

For example: \code{1970143N19091} is a storm in the North Atlantic which started on
May 23, 1970 near 19°N 91°E

See \url{http://www.ncdc.noaa.gov/ibtracs/index.php?name=numbering} for more.
}
\examples{
\donttest{
storm_data(basin='WP')
storm_data(storm='1970143N19091')
storm_data(year=1940)
storm_data(year=1941)
storm_data(year=2010)

# Or get all data, simply don't specify a value for basin, storm, or year
res <- storm_data(read=FALSE) # just get path
head()
}
}

