% Generated by roxygen2 (4.0.2): do not edit by hand
\name{storm_meta}
\alias{storm_meta}
\title{Get NOAA wind storm metadata.}
\usage{
storm_meta(what = "storm_columns")
}
\description{
Get NOAA wind storm metadata.
}
\examples{
\donttest{
head( storm_meta() )
head( storm_meta("storm_columns") )
head( storm_meta("storm_names") )
}
}

