% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tornadoes}
\alias{tornadoes}
\title{Get NOAA tornado data.}
\usage{
tornadoes(path = "~/.rnoaa/tornadoes", overwrite = TRUE)
}
\arguments{
\item{path}{A path to store the files, Default: \code{~/.ots/kelp}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}
}
\value{
A Spatial object is returned of class SpatialLinesDataFrame.
}
\description{
Get NOAA tornado data.
}
\examples{
\donttest{
shp <- tornadoes()
plot(shp) # may take 10 sec or so to render
}
}

