% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{type_summ}
\alias{type_summ}
\alias{type_summ.default}
\alias{type_summ.character}
\alias{type_summ.Date}
\alias{type_summ.factor}
\alias{type_summ.integer}
\alias{type_summ.logical}
\alias{type_summ.array}
\alias{type_summ.matrix}
\alias{type_summ.numeric}
\alias{type_summ.POSIXt}
\title{Type summary}
\usage{
type_summ(x)

\method{type_summ}{default}(x)

\method{type_summ}{character}(x)

\method{type_summ}{Date}(x)

\method{type_summ}{factor}(x)

\method{type_summ}{integer}(x)

\method{type_summ}{logical}(x)

\method{type_summ}{array}(x)

\method{type_summ}{matrix}(x)

\method{type_summ}{numeric}(x)

\method{type_summ}{POSIXt}(x)
}
\description{
Type summary
}
\keyword{internal}
