% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves.R
\name{waves}
\alias{waves}
\title{Wavewatch data}
\usage{
waves(date, grid = FALSE, ...)
}
\arguments{
\item{date}{(date/character) date in YYYY-MM format}

\item{grid}{(logical) single grid or multi, only used when date is in the overlap
years of 2005-2007 for single- and multi-grid data}

\item{...}{Curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
an object of class \code{RasterBrick}, with xx:
\itemize{
\item xx
}
}
\description{
Wavewatch data
}
\details{
The data is split into two groups:
\itemize{
\item nww3/ -- Early WW3 version, single grid from Jul 1999 - Nov 2007
with gaps
\item multi_1/ -- Multi-grid WW3, from Feb 2005 to present
}
}
\examples{
\dontrun{
waves(date = "2003-02")
waves(date = "2009-02")
}
}
\references{
ftp://polar.ncep.noaa.gov/pub/history/waves/nww3,
ftp://polar.ncep.noaa.gov/pub/history/waves/multi_1/
}
