% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-autoplot.R
\name{autoplot.meteo_coverage}
\alias{autoplot.meteo_coverage}
\title{autoplot method for meteo_coverage objects}
\usage{
\method{autoplot}{meteo_coverage}(object)
}
\arguments{
\item{object}{(data.frame) a data.frame}
}
\value{
A ggplot2 plot
}
\description{
autoplot method for meteo_coverage objects
}
\details{
see \code{\link{meteo_coverage}} for examples
}
