% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{buoy}
\alias{buoy}
\alias{buoys}
\alias{buoy_stations}
\title{Get NOAA buoy data from the National Buoy Data Center}
\usage{
buoy(dataset, buoyid, year = NULL, datatype = NULL, ...)

buoys(dataset, ...)

buoy_stations(refresh = FALSE, ...)
}
\arguments{
\item{dataset}{(character) Dataset name to query. See below for Details.
Required}

\item{buoyid}{Buoy ID, can be numeric/integer/character. Required}

\item{year}{(integer) Year of data collection. Optional}

\item{datatype}{(character) Data type, one of 'c', 'cc', 'p', 'o'. Optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}. Optional}

\item{refresh}{(logical) Whether to use cached data (\code{FALSE}) or get
new data (\code{FALSE}). Default: \code{FALSE}}
}
\description{
Get NOAA buoy data from the National Buoy Data Center
}
\details{
Functions:
\itemize{
 \item buoy_stations - Get buoy stations. A cached version of the dataset
 is available in the package. Beware, takes a long time to run if you
 do \code{refresh = TRUE}
 \item buoys - Get available buoys given a dataset name
 \item buoy - Get data given some combination of dataset name, buoy ID,
 year, and datatype
}

Options for the dataset parameter. One of:
\itemize{
 \item adcp - Acoustic Doppler Current Profiler data
 \item adcp2 - MMS Acoustic Doppler Current Profiler data
 \item cwind - Continuous Winds data
 \item dart - Deep-ocean Assessment and Reporting of Tsunamis data
 \item mmbcur - Marsh-McBirney Current Measurements data
 \item ocean - Oceanographic data
 \item pwind - Peak Winds data
 \item stdmet - Standard Meteorological data
 \item swden - Spectral Wave Density data with Spectral Wave Direction data
 \item wlevel - Water Level data
}
}
\examples{
\dontrun{
# Get buoy station information
x <- buoy_stations()
library("leaflet")
leaflet(data = na.omit(x)) \%>\%
  leaflet::addTiles() \%>\%
  leaflet::addCircles(~lon, ~lat, opacity = 0.5)

# Get available buoys
buoys(dataset = 'cwind')

# Get data for a buoy
## if no year or datatype specified, we get the first file
buoy(dataset = 'cwind', buoyid = 46085)

# Including specific year
buoy(dataset = 'cwind', buoyid = 41001, year = 1999)

# Including specific year and datatype
buoy(dataset = 'cwind', buoyid = 41001, year = 2008, datatype = "cc")
buoy(dataset = 'cwind', buoyid = 41001, year = 2008, datatype = "cc")

# Other datasets
buoy(dataset = 'ocean', buoyid = 41029)

# curl debugging
library('httr')
buoy(dataset = 'cwind', buoyid = 46085, config=verbose())

# some buoy ids are character, case doesn't matter, we'll account for it
buoy(dataset = "stdmet", buoyid = "VCAF1")
buoy(dataset = "stdmet", buoyid = "wplf1")
buoy(dataset = "dart", buoyid = "dartu")
}
}
\references{
\url{http://www.ndbc.noaa.gov/}, \url{http://dods.ndbc.noaa.gov/}
}
