% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{caching}
\alias{caching}
\alias{ghcnd_clear_cache}
\title{Clear cached files}
\usage{
ghcnd_clear_cache(path = "~/.rnoaa/ghcnd", force = FALSE)
}
\arguments{
\item{path}{Path to location of cached files. Defaults to \code{disk()$path}}

\item{force}{(logical) Should we force removal of files if permissions say otherwise?}
}
\description{
Clear cached files
}
\details{
BEWARE: this will clear all cached files.
}

