% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{check_response_swdi}
\alias{check_response_swdi}
\title{Check response from NOAA SWDI service, including status codes, server error messages,
mime-type, etc.}
\usage{
check_response_swdi(x, format)
}
\description{
Check response from NOAA SWDI service, including status codes, server error messages,
mime-type, etc.
}
\keyword{internal}
