% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{clean_daily}
\alias{clean_daily}
\title{Clean the daily summary data}
\usage{
clean_daily(stationid, keep_flags = FALSE, var = "all", date_min = NULL,
  date_max = NULL)
}
\arguments{
\item{stationid}{A character vector giving the identification of the weather
station for which the user would like to pull data. To get a full and
current list of stations, the user can use the \code{\link{ghcnd_stations}}
function.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all the flags
for each weather varialbe. The default is to not keep the flags (FALSE).}

\item{var}{A character vector specifying either \code{"all"} (pull all
available weather parameters for the site) or the weather parameters to
keep in the final data (e.g., \code{c("TMAX", "TMIN")} to only keep
maximum and minimum temperature). Example choices for this argument include:
\itemize{
\item \code{PRCP}: Precipitation, in tenths of millimeters
\item \code{TAVG}: Average temperature, in tenths of degrees Celsius
\item \code{TMAX}: Maximum temperature, in tenths of degrees Celsius
\item \code{TMIN}: Minimum temperature, in tenths of degrees Celsius
}
A full list of possible weather variables is available in NOAA's README
file for the GHCND data (\url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}).
Most weather stations will only have a small subset of all the possible
weather variables, so the data generated by this function may not include
all of the variables the user specifies through this argument.}

\item{date_min}{Character strings giving, respectively, the earliest
and latest dates of the daily weather time series that the user would
like in the final output. This character string should be formatted as
"yyyy-mm-dd" (see examples). If these arguments are not specified, the
default is to keep all available daily data for the queried weather site.}

\item{date_max}{Character strings giving, respectively, the earliest
and latest dates of the daily weather time series that the user would
like in the final output. This character string should be formatted as
"yyyy-mm-dd" (see examples). If these arguments are not specified, the
default is to keep all available daily data for the queried weather site.}

\item{ghcnd_data}{An object of class "ghcnd", as generated for a single weather
station using \code{\link{ghcnd}}.}
}
\value{
A data frame of daily weather data for a single weather monitor,
   converted to a tidy format. All weather variables may not exist for all
   weather stations, but the returned dataset would have at most the
   following columns:
   \itemize{
   \item \code{id}: Character string with the weather station site id
   \item \code{date}: Date of the observation
   \item \code{prcp}: Precipitation, in mm
   \item \code{tavg}: Average temperature, in degrees Celsius
   \item \code{tmax}: Maximum temperature, in degrees Celsius
   \item \code{tmin}: Minimum temperature, in degrees Celsius
   \item \code{awnd}: Average daily wind speed, in meters / second
   \item \code{wsfg}: Peak gust wind speed, in meters / second
   }
   There are other possible weather variables in the Global Historical
   Climatology Network, but we have not implemented cleaning them through
   this function. See
   \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt} for a full
   list. If you are interested in having other weather variables from this
   list added to this function, you can submit a request to this package's
   GitHub issues page and we may be able to add it.
}
\description{
This function inputs an object created by \code{\link{ghcnd}} and cleans up the
data into a tidy form.
}
\details{
The following reference has more information on this data:

Menne, M.J., I. Durre, R.S. Vose, B.E. Gleason, and T.G. Houston, 2012:
An overview of the Global Historical Climatology Network-Daily Database.
Journal of Atmospheric and Oceanic Technology, 29, 897-910,
doi:10.1175/JTECH-D-11-00103.1.
}
\note{
The weather flags, which are kept by specifying
\code{keep_flags = TRUE} are:
\itemize{
\item \code{*_mflag}: Measurement flag, which gives some information on how
   the observation was measured.
\item \code{*_qflag}: Quality flag, which gives quality information on the
   measurement, like if it failed to pass certain quality checks.
\item \code{*_sflag}: Source flag. This gives some information on the
   weather collection system (e.g., U.S. Cooperative Summary of the Day,
   Australian Bureau of Meteorology) the weather observation comes from.
}
More information on the interpretation of these flags can be found in the
README file for the NCDC's Daily Global Historical Climatology Network's
data at \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}.
}
\examples{
\dontrun{
# One station in Australia is ASM00094275
cleaned_df <- clean_daily(stationid = "ASN00003003")
}

}

