% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{clean_daily}
\alias{clean_daily}
\title{Clean the daily summary data}
\usage{
clean_daily(ghcnd_data, keep_flags = FALSE)
}
\arguments{
\item{ghcnd_data}{An object of class "ghcnd", as generated for a single weather
station using \code{\link{ghcnd}}.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all the flags
for each weather varialbe. The default is to not keep the flags (FALSE).}
}
\description{
This function inputs an object created by \code{\link{ghcnd}} and cleans up the
data into a tidy form.
}
\examples{
\dontrun{
# You must have your NOAA API key saved as `steve`
options("noaakey" = steve)

# One station in Australia is ASM00094275
ghcnd_data <- ghcnd(stationid = "ASN00003003")
cleaned_df <- clean_daily(ghcnd_data)
}

}

