% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/co-ops.R
\name{coops}
\alias{coops}
\alias{coops_search}
\title{Get NOAA co-ops data}
\usage{
coops_search(begin_date = NULL, end_date = NULL, station_name = NULL,
  product = NULL, datum = NULL, units = "metric", time_zone = "gmt",
  application = "web_services", format = "json")
}
\arguments{
\item{begin_date}{numeric date in yyyymmdd format}

\item{end_date}{numeric date in yyyymmdd format}

\item{station_name}{numeric}

\item{product}{character choice of water_level, air_temperature, water_temperature, wind, air_pressure, air_gap, conductivity, visibility, humidity, salinity, hourly_height, high_low, daily_mean, monthly_mean, one_minute_water_level, predictions, datums, currents.}

\item{datum}{character choice of MHHW, MHW, MTL, MSL, MLW, MLLW, NAVD, STND}
}
\description{
Get NOAA co-ops data
}
\examples{
\dontrun{
coops_search(station_name = 8723970, begin_date = 19820301, end_date = 20141001, datum = "stnd", product = "monthly_mean")
}
}
\references{
\url{http://co-ops.nos.noaa.gov/api/}
}

