% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates the R Leaflet maps' labels with the quake's traits.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
object with the earthquake traits(i.e., location name, magnitude, and
total deaths).}
}
\description{
Creates the R Leaflet maps' labels with the quake's traits.
}
