% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the location name of the earthquake.}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A data frame or tibble object with the above variable.}
}
\value{
Returns a \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
object.
}
\description{
With \code{eq_location_clean} you can get rid of the country in the \code{LOCATION_NAME}
variable of the \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{NOAA's Significant Earthquake database}
and convert it into title case.
}
\examples{
raw_data <- download_earthquake_data()
clean_data <- eq_location_clean(raw_data)
}
\seealso{
\code{\link{eq_clean_data}} for process other key variables.
}
