% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-functions.R
\name{eq_map}
\alias{eq_map}
\title{Displays an R Leaflet map with the quake's epicenters.}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
object with the earthquake epicenters.}
}
\description{
Displays an R Leaflet map with the quake's epicenters.
}
