% Generated by roxygen2 (4.0.1): do not edit by hand
\name{erddap_search}
\alias{erddap_search}
\alias{print.erddap_search}
\title{Search for ERDDAP datasets.}
\usage{
erddap_search(query, page = NULL, page_size = NULL, callopts = list())

\method{print}{erddap_search}(x, ...)
}
\arguments{
\item{query}{Search terms}

\item{page}{Page number}

\item{page_size}{Results per page}

\item{callopts}{Further args passed on to httr::GET (must be a named parameter)}

\item{x}{Input to print method for class erddap_search}

\item{...}{Further args to print, ignored.}
}
\description{
Search for ERDDAP datasets.
}
\examples{
\dontrun{
(out <- erddap_search(query='fish size'))
out$alldata
(out <- erddap_search(query='size'))
out$info
}
}

