% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnd.R
\name{ghcnd_splitvars}
\alias{ghcnd_splitvars}
\title{Split variables in data returned from \code{ghcnd}}
\usage{
ghcnd_splitvars(x)
}
\arguments{
\item{x}{An object returned from \code{\link[=ghcnd]{ghcnd()}}}
}
\description{
This function is a helper function for \code{\link[=ghcnd_search]{ghcnd_search()}}. It helps
with cleaning up the data returned from \code{\link[=ghcnd]{ghcnd()}}, to get it in a
format that is easier to work with.
}
\note{
See \code{\link[=ghcnd]{ghcnd()}} examples
}
\author{
Scott Chamberlain, Adam Erickson, Elio Campitelli
}
