% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/homr_definitions.R
\name{homr_definitions}
\alias{homr_definitions}
\title{Historical Observing Metadata Repository (HOMR) station metadata - definitions}
\usage{
homr_definitions(...)
}
\arguments{
\item{...}{Named parameters, such as \code{query}, \code{path}, etc, passed on to
\code{\link[httr]{modify_url}}. Unnamed parameters will be combined with
\code{\link[httr]{config}}.}
}
\description{
Historical Observing Metadata Repository (HOMR) station metadata - definitions
}
\examples{
\dontrun{
head( homr_definitions() )
}
}

