% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homr_definitions.R
\name{homr_definitions}
\alias{homr_definitions}
\title{Historical Observing Metadata Repository (HOMR) station metadata -
definitions}
\usage{
homr_definitions(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}} (optional)}
}
\description{
Historical Observing Metadata Repository (HOMR) station metadata -
definitions
}
\examples{
\dontrun{
head( homr_definitions() )
}
}
