% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_nearby_stations}
\alias{meteo_nearby_stations}
\title{get_nearby_stations}
\usage{
meteo_nearby_stations(lat_lon_df, lat_colname, lon_colname, ghcnd_station_list,
  radius, ...)
}
\arguments{
\item{lat_lon_df}{A dataframe that contains the site latitude and longitude values used to search for nearby weather stations}

\item{lat_colname}{Name of the latitude column in the lat_lon_df}

\item{lon_colname}{Name of the longitude column in the lat_lon_df}

\item{ghcnd_station_list}{List of weather stations obtained using `ghcnd_stations()`}

\item{radius}{Radius to search (in km)}
}
\value{
a dataframe containing a unique set of the weather stations within the search radius
}
\description{

}
\examples{
lat <- c(37.779199, 37.531635)
lon <- c(-122.404294, -122.419282)
lat_lon_df <- data.frame(latitude = lat, longitude = lon)

nearby_stations <- 
get_nearby_stations(lat_lon_df = test_df, lat_colname = 'latitude', lon_colname = 'longitude', ghcnd_station_list = stations, radius = 20)

}

