% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_nearby_stations}
\alias{meteo_nearby_stations}
\title{Find weather monitors near locations}
\usage{
meteo_nearby_stations(lat_lon_df, lat_colname = "latitude",
  lon_colname = "longitude", station_data = ghcnd_stations()[[1]],
  var = "all", year_min = NULL, year_max = NULL, radius = 25, ...)
}
\arguments{
\item{lat_lon_df}{A dataframe that contains the site latitude and longitude
values used to search for nearby weather stations}

\item{lat_colname}{A character string giving the name of the latitude column
in the \code{lat_lon_df} dataframe.}

\item{lon_colname}{A character string giving the name of the longitude column
in the \code{lat_lon_df} dataframe.}

\item{station_data}{The output of \code{ghcnd_stations()[[1]]}:
a current list of weather stations available through NOAA for the GHCND
dataset. The format of this is a list with a single element, a dataframe
with one row per available weather station. To save time, run the
\code{ghcnd_stations} call and save the output to an object, rather than
rerunning the default every time (see the examples).}

\item{var}{A character vector specifying either \code{"all"} (pull all
available weather parameters for the site) or the weather parameters to
keep in the final data (e.g., \code{c("TMAX", "TMIN")} to only keep
maximum and minimum temperature). Example choices for this argument include:
\itemize{
\item \code{PRCP}: Precipitation, in tenths of millimeters
\item \code{TAVG}: Average temperature, in tenths of degrees Celsius
\item \code{TMAX}: Maximum temperature, in tenths of degrees Celsius
\item \code{TMIN}: Minimum temperature, in tenths of degrees Celsius
}
A full list of possible weather variables is available in NOAA's README
file for the GHCND data (\url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}).
Most weather stations will only have a small subset of all the possible
weather variables, so the data generated by this function may not include
all of the variables the user specifies through this argument.}

\item{radius}{A numeric vector giving the radius (in kilometers) within which
to search for monitors near a location}

\item{...}{Additional curl options to pass through to \code{\link[httr]{GET}}.}
}
\value{
a dataframe containing a unique set of the weather stations within
   the search radius. Site IDs for the weather stations given in this
   dataframe can be used in conjunction with other functions in the
   \code{rnoaa} package to pull weather data for the station.
}
\description{
This function inputs a dataframe with latitudes and longitudes of locations
and creates a dataframe with monitors with a certain radius of those
locations. The site ids from this dataframe can then be used with other
\code{rnoaa} functions to pull data from all available weather stations near
a location.
}
\note{
By default, this function will pull the full station list from NOAA
   to use to identify nearby locations. If you will be creating lists of
   monitors nearby several stations, you can save some time by using the
   \code{\link{ghcnd_stations}} function separately to create an object
   with all stations and then use the argument \code{ghcnd_station_list} in
   this function to reference that object, rather than using this function's
   defaults.
}
\examples{
\dontrun{
lat <- c(37.779199, 37.531635)
lon <- c(-122.404294, -122.419282)
lat_lon_df <- data.frame(latitude = lat, longitude = lon)

station_data <- ghcnd_stations()[[1]] # Takes a while to run
nearby_stations <-  meteo_nearby_stations(lat_lon_df,
                    station_data = station_data, radius = 50)

miami <- data.frame(id = "miami", latitude = 25.7617, longitude = -80.1918)
meteo_nearby_stations(lat_lon_df = miami, station_data = station_data,
                      radius = 50, var = c("prcp", "tmax"),
                      year_min = 1992, year_max = 1992)
}

}

