% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_process_geographic_data}
\alias{meteo_process_geographic_data}
\title{meteo_process_geographic_data}
\usage{
meteo_process_geographic_data(data, lat, long, latName, longName,
  units = "deg")
}
\arguments{
\item{data}{data}

\item{lat}{data}

\item{long}{data}

\item{latName}{data}

\item{longName}{data}

\item{units}{data}
}
\description{
meteo_process_geographic_data
}

