% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_process_geographic_data}
\alias{meteo_process_geographic_data}
\title{meteo_process_geographic_data}
\usage{
meteo_process_geographic_data(data, lat, long, latName, longName,
  units = "deg")
}
\arguments{
\item{lat}{Latitude to centre search at}

\item{long}{Longitude to centre search at}

\item{latName}{Name of latitude header name in data, Default = 'latitude'}

\item{longName}{Name of longitude header name in data. Default = 'longitude'}

\item{units}{Units of the latitude and longitude values: degrees 'deg',
radians 'rad', d/m/s 'dms'. Default = 'deg'}
}
\description{
meteo_process_geographic_data
}

