% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{meteo_pull_monitors}
\alias{meteo_pull_monitors}
\title{Pull daily weather data for multiple weather monitors}
\usage{
meteo_pull_monitors(monitors, keep_flags = FALSE)
}
\arguments{
\item{monitors}{A character vector listing the station IDs for all
weather stations the user would like to pull.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all the flags
for each weather varialbe. The default is to not keep the flags (FALSE).}
}
\value{
A data frame of daily weather data for a multiple weather monitors,
   converted to a tidy format. All weather variables may not exist for all
   weather stations, but the returned dataset would have at most the
   following columns:
   \itemize{
   \item \code{id}: Character string with the weather station site id
   \item \code{date}: Date of the observation
   \item \code{prcp}: Precipitation, in mm
   \item \code{tavg}: Average temperature, in degrees Celsius
   \item \code{tmax}: Maximum temperature, in degrees Celsius
   \item \code{tmin}: Minimum temperature, in degrees Celsius
   \item \code{awnd}: Average daily wind speed, in meters / second
   \item \code{wsfg}: Peak gust wind speed, in meters / second
   }
   There are other possible weather variables in the Global Historical
   Climatology Network, but we have not implemented cleaning them through
   this function. See
   \url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt} for a full
   list. If you are interested in having other weather variables from this
   list added to this function, you can submit a request to this package's
   GitHub issues page and we may be able to add it.
}
\description{
This function takes a vector of one or more weather station IDs. It will pull
the weather data from the Global Historical Climatology Network's daily
data for each of the stations and join them together in a single tidy
dataframe.
}
\note{
This function may take a while to run.
}
\examples{
\dontrun{

monitors <- c("ASN00003003", "ASM00094299", "ASM00094995", "ASM00094998")
all_monitors_clean <- meteo_pull_monitors(monitors)

}

}

