% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ghcnd.R
\name{meteo_pull_monitors}
\alias{meteo_pull_monitors}
\title{Pull daily weather data for multiple weather monitors}
\usage{
meteo_pull_monitors(monitors, keep_flags = FALSE)
}
\arguments{
\item{monitors}{A character vector listing the station IDs for all
weather stations the user would like to pull.}

\item{keep_flags}{TRUE / FALSE for whether the user would like to keep all the flags
for each weather varialbe. The default is to not keep the flags (FALSE).}
}
\description{
This function takes a vector of one or more weather station IDs. It will pull
the weather data from the Global Historical Climatology Network's daily
data for each of the stations and join them together in a single tidy
dataframe.
}
\examples{
\dontrun{

monitors <- c("ASN00003003", "ASM00094299", "ASM00094995", "ASM00094998")
all_monitors_clean <- meteo_pull_monitors(monitors)
}
}

