% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_distance.R
\name{meteo_spherical_distance}
\alias{meteo_spherical_distance}
\title{meteo_spherical_distance}
\usage{
meteo_spherical_distance(lat1, long1, lat2, long2, units = "deg")
}
\arguments{
\item{lat1}{data}

\item{long1}{data}

\item{lat2}{data}

\item{long2}{data}

\item{units}{data}
}
\description{
meteo_spherical_distance
}

