% Generated by roxygen2 (4.0.0): do not edit by hand
\name{noaa}
\alias{noaa}
\title{Get NOAA data.}
\usage{
noaa(datasetid = NULL, datatypeid = NULL, stationid = NULL,
  locationid = NULL, startdate = NULL, enddate = NULL, sortfield = NULL,
  sortorder = NULL, limit = 25, offset = NULL, callopts = list(),
  token = NULL, dataset = NULL, datatype = NULL, station = NULL,
  location = NULL, locationtype = NULL, page = NULL, year = NULL,
  month = NULL, day = NULL, results = NULL)
}
\arguments{
  \item{datatypeid}{Accepts a valid data type id or a chain
  of data type ids in a comma-separated vector. Data
  returned will contain all of the data type(s) specified
  (optional)}

  \item{locationid}{Accepts a valid location id or a chain
  of location ids in a comma-separated vector. Data
  returned will contain data for the location(s) specified
  (optional)}

  \item{stationid}{Accepts a valid station id or a chain of
  of station ids in a comma-separated vector. Data returned
  will contain data for the station(s) specified
  (optional)}

  \item{sortfield}{The field to sort results by. Supports
  id, name, mindate, maxdate, and datacoverage fields
  (optional)}

  \item{sortorder}{Which order to sort by, asc or desc.
  Defaults to asc (optional)}

  \item{limit}{Defaults to 25, limits the number of results
  in the response. Maximum is 1000 (optional)}

  \item{offset}{Defaults to 0, used to offset the
  resultlist (optional)}

  \item{token}{This must be a valid token token supplied to
  you by NCDC's Climate Data Online access token generator.
  (required) Get an API key (=token) at
  \url{http://www.ncdc.noaa.gov/cdo-web/token}. You can
  pass your token in as an argument or store it in your
  .Rprofile file with an entry like \itemize{ \item
  options("noaakey" = "your-noaa-token") }}

  \item{callopts}{Further arguments passed on to the API
  GET call. (optional)}

  \item{datasetid}{(required) Accepts a single valid
  dataset id. Data returned will be from the dataset
  specified, see datasets()}

  \item{startdate}{(required) Accepts valid ISO formated
  date (yyyy-mm-dd). Data returned will have data after the
  specified date. Paramater can be use independently of
  enddate}

  \item{enddate}{(required) Accepts valid ISO formated date
  (yyyy-mm-dd). Data returned will have data before the
  specified date. Paramater can be use independently of
  startdate}

  \item{dataset}{THIS IS A DEPRECATED ARGUMENT. See
  datasetid.}

  \item{datatype}{THIS IS A DEPRECATED ARGUMENT. See
  datatypeid.}

  \item{station}{THIS IS A DEPRECATED ARGUMENT. See
  stationid.}

  \item{location}{THIS IS A DEPRECATED ARGUMENT. See
  locationid.}

  \item{locationtype}{THIS IS A DEPRECATED ARGUMENT. There
  is no equivalent argument in v2 of the NOAA API.}

  \item{page}{THIS IS A DEPRECATED ARGUMENT. There is no
  equivalent argument in v2 of the NOAA API.}

  \item{year}{THIS IS A DEPRECATED ARGUMENT. Use
  combination of startdate and enddate arguments.}

  \item{month}{THIS IS A DEPRECATED ARGUMENT. Use
  combination of startdate and enddate arguments.}

  \item{day}{THIS IS A DEPRECATED ARGUMENT. Use combination
  of startdate and enddate arguments.}

  \item{results}{THIS IS A DEPRECATED ARGUMENT. See limit.}
}
\value{
An S3 list of length two, a slot of metadata (meta), and a slot for data (data).
The meta slot is a list of metadata elements, and the data slot is a data.frame,
possibly of length zero if no data is found.
}
\description{
Get NOAA data.
}
\details{
Note that NOAA API calls can take a long time depending on the call. The NOAA API doesn't
perform well with very long timespans, and will time out and make you angry - beware.

Keep in mind that three parameters, datasetid, startdate, and enddate are required.

Note that the default limit (no. records returned) is 25.
}
\examples{
\dontrun{
# GHCN-Daily (or GHCND) data, for a specific station
noaa(datasetid='GHCND', stationid='GHCND:USW00014895', startdate = '2013-10-01', enddate = '2013-12-01')

# GHCND data, for a location by FIPS code
noaa(datasetid='GHCND', locationid = 'FIPS:02', startdate = '2010-05-01',
   enddate = '2010-05-10')

# GHCND data from October 1 2013 to December 1 2013
noaa(datasetid='GHCND', startdate = '2013-10-01', enddate = '2013-12-01')

# GHCN-Monthly (or GHCNDMS) data from October 1 2013 to December 1 2013
noaa(datasetid='GHCNDMS', startdate = '2013-10-01', enddate = '2013-12-01')

# Normals Daily (or NORMAL_DLY) GHCND:USW00014895 dly-tmax-normal data
noaa(datasetid='NORMAL_DLY', stationid='GHCND:USW00014895', startdate = '2010-05-01',
   enddate = '2010-05-10')

# Dataset, and location in Australia
noaa(datasetid='GHCND', locationid='FIPS:AS', startdate = '2010-05-01', enddate = '2010-05-31')

# Dataset, location and datatype for PRECIP_HLY
noaa(datasetid='PRECIP_HLY', locationid='ZIP:28801', datatypeid='HPCP', startdate = '2010-05-01',
   enddate = '2010-05-10')

# Dataset, location, station and datatype
noaa(datasetid='PRECIP_HLY', locationid='ZIP:28801', stationid='COOP:310301', datatypeid='HPCP',
   startdate = '2010-05-01', enddate = '2010-05-10')

# Dataset, location, and datatype for GHCND
noaa(datasetid='GHCND', locationid='FIPS:BR', datatypeid='PRCP', startdate = '2010-05-01',
   enddate = '2010-05-10')

# Normals Daily GHCND dly-tmax-normal data
noaa(datasetid='NORMAL_DLY', datatypeid='dly-tmax-normal', startdate = '2010-05-01',
   enddate = '2010-05-10')

# Normals Daily GHCND:USW00014895 dly-tmax-normal
noaa(datasetid='NORMAL_DLY', stationid='GHCND:USW00014895', datatypeid='dly-tmax-normal',
   startdate = '2010-05-01', enddate = '2010-05-10')

# Hourly Precipitation data for ZIP code 28801
noaa(datasetid='PRECIP_HLY', locationid='ZIP:28801', datatypeid='HPCP', startdate = '2010-05-01',
   enddate = '2010-05-10')

# Search the NORMAL_HLY dataset
noaa(datasetid='NORMAL_HLY', stationid = 'GHCND:USW00003812', startdate = '2010-05-01', enddate = '2010-05-10')

# Search the ANNUAL dataset
noaa(datasetid='ANNUAL', locationid='ZIP:28801', startdate = '2010-05-01', enddate = '2010-05-10')
}

\donttest{
# NEXRAD2 data
## doesn't work yet
noaa(datasetid='NEXRAD2', startdate = '2013-10-01', enddate = '2013-12-01')
}
}

