% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readshpfile}
\alias{readshpfile}
\title{Function to read shapefiles}
\usage{
readshpfile(x, storepath = NULL)
}
\arguments{
\item{x}{A url}

\item{storepath}{Path to store data in}
}
\value{
An object of class sp
}
\description{
Function to read shapefiles
}
\keyword{internal}

