\name{readshpfile}
\alias{readshpfile}
\title{Function to read shapefile and}
\usage{
  readshpfile(x, storepath = "~/seaicedata")
}
\arguments{
  \item{x}{A url}

  \item{storepath}{Path to store data in}
}
\value{
  An object of class sp
}
\description{
  Function to read shapefile and
}
\keyword{internal}

