% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rnoaa-defunct}
\alias{rnoaa-defunct}
\title{Defunct functions in rnoaa}
\description{
\itemize{
\item \code{\link{noaa}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_datacats}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_datasets}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_datatypes}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_locs}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_locs_cats}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_stations}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_plot}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_combine}}: Function name changed, prefixed with ncdc now
\item \code{\link{noaa_seaice}}: Function name changed to seaice
\item \code{\link{erddap_data}}: See package rerddap
\item \code{\link{erddap_clear_cache}}: See package rerddap
\item \code{\link{erddap_datasets}}: Moved to \code{ed_datasets} in
package rerddap
\item \code{\link{erddap_grid}}: Moved to \code{griddap} in package
rerddap
\item \code{\link{erddap_info}}: Moved to \code{info} in package
rerddap
\item \code{\link{erddap_search}}: Moved to \code{ed_search} in
package rerddap
\item \code{\link{erddap_table}}: Moved to \code{tabledap} in
package rerddap
\item \code{\link{ncdc_leg_variables}}: Removed. See \code{NCDC Legacy} below
\item \code{\link{ncdc_leg_sites}}: Removed. See \code{NCDC Legacy} below
\item \code{\link{ncdc_leg_site_info}}: Removed. See \code{NCDC Legacy} below
\item \code{\link{ncdc_leg_data}}: Removed. See \code{NCDC Legacy} below
\item \code{\link{seaice}}: Replaced with \code{\link{sea_ice}}
\item \code{\link{lcd_cleanup}}: No longer available. See \code{\link{lcd}}
docs
\item \code{\link{ghcnd_clear_cache}}: No longer available.
See \code{\link{rnoaa_caching}}
}
}
\details{
The functions for working with GEFS ensemble forecast data (prefixed with
"gefs") are defunct, but may come back to rnoaa later. They are:

\itemize{
\item \code{\link{gefs}}
\item \code{\link{gefs_dimension_values}}
\item \code{\link{gefs_dimensions}}
\item \code{\link{gefs_ensembles}}
\item \code{\link{gefs_latitudes}}
\item \code{\link{gefs_longitudes}}
\item \code{\link{gefs_times}}
\item \code{\link{gefs_variables}}
}
}
\section{NCDC Legacy}{

The NCDC legacy API is too unreliable and slow. Use the newer NCDC API via the
functions \code{\link{ncdc}}, \code{\link{ncdc_datacats}}, \code{\link{ncdc_datasets}},
\code{\link{ncdc_datatypes}}, \code{\link{ncdc_locs}}, \code{\link{ncdc_locs_cats}},
\code{\link{ncdc_stations}}, \code{\link{ncdc_plot}}, and \code{\link{ncdc_combine}}
}

