% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{rnoaa-defunct}
\alias{rnoaa-defunct}
\title{Defunct functions in rnoaa}
\description{
\itemize{
 \item \code{\link{noaa}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datacats}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datasets}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_datatypes}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_locs}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_locs_cats}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_stations}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_plot}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_combine}}: Function name changed, prefixed with ncdc now
 \item \code{\link{noaa_seaice}}: Function name changed to seaice
 \item \code{\link{erddap_data}}: See the function \code{\link{erddap_grid}} for getting
 griddap data and \code{\link{erddap_table}} for getting tabledap data.
}
}

