% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rnoaa-package}
\alias{rnoaa}
\alias{rnoaa-package}
\title{General purpose R interface to noaa.}
\description{
rnoaa is an R interface to NOAA climate data.
}
\details{
Specifically, most functions in this package interact
with the National Climatic Data Center application
programming interface (API) at
\url{http://www.ncdc.noaa.gov/cdo-web/webservices/v2}.

An access token, or API key, is required to use this
R package. The key is required by NOAA, not the
creators of this R package. Go to the link given above
to get an API key.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

