% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnoaa-package.r
\docType{data}
\name{storm_columns}
\alias{storm_columns}
\title{NOAA storm column descriptions for data from IBTrACS}
\format{A data frame with 195 rows and 8 variables}
\description{
This dataset includes description of the columns of each dataset acquired
using \code{\link[rnoaa]{storm_data}}
}
\keyword{datasets}
