% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.R, R/storms_meta.R
\name{storm_data}
\alias{storm_data}
\alias{storm_meta}
\title{Get NOAA wind storm tabular data, metadata, or shp files from IBTrACS}
\usage{
storm_data(...)

storm_meta(...)
}
\description{
Get NOAA wind storm tabular data, metadata, or shp files from IBTrACS

storm_meta
}
\keyword{internal}
