% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnoaa-package.r
\docType{data}
\name{storm_names}
\alias{storm_names}
\title{NOAA storm names from IBTrACS}
\format{A data frame with 12,209 rows and 2 variables}
\description{
This dataset includes a crosswalk from storm serial numbers to their names.
Storm serial numbers are used to search for storms in the
\code{\link[rnoaa]{storm_data}} function.
}
\keyword{datasets}
