% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornadoes.R
\name{tornadoes}
\alias{tornadoes}
\title{Get NOAA tornado data.}
\usage{
tornadoes(path = "~/.rnoaa/tornadoes", overwrite = TRUE, ...)
}
\arguments{
\item{path}{A path to store the files, Default: \code{~/.ots/kelp}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} (optional)}
}
\value{
A Spatial object is returned of class SpatialLinesDataFrame.
}
\description{
Get NOAA tornado data.
}
\examples{
\dontrun{
shp <- tornadoes()
library('sp')
plot(shp) # may take 10 sec or so to render
}
}
\references{
\url{http://www.spc.noaa.gov/gis/svrgis/}
}

