% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornadoes.R
\name{tornadoes}
\alias{tornadoes}
\title{Get NOAA tornado data.}
\usage{
tornadoes(...)
}
\arguments{
\item{...}{Curl options passed on to \link[crul:verb-GET]{crul::verb-GET} (optional)}
}
\value{
A Spatial object is returned of class SpatialLinesDataFrame.
}
\description{
This function gets spatial paths of tornadoes from NOAA's National Weather
Service Storm Prediction Center Severe Weather GIS web page.
}
\note{
See \link{torn_cache} for managing cached files
}
\examples{
\dontrun{
shp <- tornadoes()
library('sp')
if (interactive()) {
  # may take 10 sec or so to render
  plot(shp)
}
}
}
\references{
https://www.spc.noaa.gov/gis/svrgis/
}
